+++ CHANGELOG.md
@@ -20,6 +20,8 @@ the detailed section referring to by linking pull requests or issues.
 
 #### Patch
 
+- Data Offer details now display the contract ID for each contract offer
+
 ## [v4.1.3] - 2024-09-03
 
 ### Overview